+++ modules/data.remote/R/Prep_AGB_IC_from_2010_global.R
@@ -20,24 +20,27 @@ Prep_AGB_IC_from_2010_global <- function(site_info, paths.list, ens) {
     future::plan(future::multisession)
   }
   ## get coordinates and provide spatial info
-  site_coords <- data.frame(site_info$lon, site_info$lat)
-  names(site_coords) <- c("Longitude","Latitude")
-  coords_latlong <- sp::SpatialPoints(site_coords)
-  sp::proj4string(coords_latlong) <- sp::CRS("+init=epsg:4326")
+  coords_latlong <- terra::vect(cbind(site_info$lon, site_info$lat), 
+                                crs = "+proj=longlat +datum=WGS84 +no_defs")
   ## load gridded AGB data
-  raster_data <- lapply(paths.list, raster::raster)
+  raster_data <- terra::rast(unlist(paths.list))
   ## reproject Lat/Long site coords to AGB Albers Equal-Area
-  coords_AEA <- sp::spTransform(coords_latlong,
-                                raster::crs(raster::raster(raster_data[[1]])))
-  ## prepare product for extraction - stack requested years
-  raster_data_stack <- raster::stack(raster_data)
+  coords_AEA <- terra::project(coords_latlong, terra::crs(raster_data))
   ## extract
-  agb_pixel <- raster::extract(x = raster_data_stack, 
-                               y = coords_AEA, buffer=0, fun=NULL, df=FALSE)
-  sampled_ic <- agb_pixel %>% furrr::future_map(function(pixel){
-    ens_sample <- stats::rnorm(ens, pixel["mean"], pixel["uncertainty"])
-    ens_sample[which(ens_sample<0)] <- 0
-    ens_sample
-  }, .progress = T) %>% dplyr::bind_cols() %>% `colnames<-`(site_info$site_id)
+  agb_pixel <- terra::extract(x = raster_data, y = coords_AEA) %>%
+    `colnames<-`(c("site.id", "mean", "uncertainty"))
+  sampled_ic <- split(as.data.frame(agb_pixel[, c(2,3)]), 
+                      seq(nrow(as.data.frame(agb_pixel[, c(2,3)])))) %>% 
+    furrr::future_map(function(pixel){
+      # make sure there is never a zero uncertainty.
+      if (pixel[,"uncertainty"] == 0) {
+        uncertainty <- .1
+      } else {
+        uncertainty <- pixel[,"uncertainty"]
+      }
+      ens_sample <- stats::rnorm(n = ens, mean = pixel[,"mean"], sd = uncertainty)
+      ens_sample[which(ens_sample<0)] <- 0
+      ens_sample
+    }, .progress = T) %>% dplyr::bind_cols() %>% `colnames<-`(site_info$site_id)
   return(sampled_ic)
 }