+++ CHANGELOG.md
@@ -64,6 +64,7 @@ For more information about this file see also [Keep a Changelog](http://keepacha
 - Added new feature of downloading datasets from the NASA DAAC ORNL database.
 - Extended downscale function and created 'downscale_hrly' so that it handles more frequent data
 - Added 'aggregate' as a new feature for downscaled data
+- Added ERA5 download function that applies to the new CDS API.
 
 ### Fixed